Updated Stylesheets
The Stylesheet generetor is approaching supporting 100 different screen resolutions on a multitude of devices. The plan is to maintain the table, then copy/ past the output (via GIT) to the website, where the interactive demo will update. Devices on the network can the retrieve the new stylesheets from the RAW.GITHUB as part of the OS. Since the interactive demo, website, dashboard and mode select menu will all use these stylesheets.