interactive demo
temp resolution for interactive demo , until the breakpoints of the UI are sorted. Because the UI looks at the devices screen size, not the iframe it's in. And when it doesn't match a supported devices it has no options. But when the UI breakpoint is sorted it'll have options. Then I can just set the lightbox to 80% width  (and maybe use breakpoints for the height)
 and the rest will figure itself out.