solution for distributor registration
so the applicant will submit a google form. the data will be timestamped and so will the thank you page. this should match so it can serve as a uid. What they write down from the thank you page should match the datestamp we have on record in the spreadsheet. and since it's by the second, it should be unuiqe. unless a freak scenario happens whereby two register the exact same second, but unlikely and we'll hit that bridge when we come to it