fixed demo stylesheet
the previous solution developed with Roy got way complicated.
So I reverted back to a basic viewport solution.
There was a reason why we devised the complex solution, perhaps it might still need to be used for the copy of the UI which runs on the device. This is done by simply changing the stylesheets from -frames-L2 to -frames-L.css - but hopefully wont need to .