interactive demo/ UI & technical docs
The tech doc library rebuilds itself now and re-builds the PDF's by running make.sh (command is 'sh make.sh'). Inside the make.sh script file there's a few things going on, open it to see. Theme changes, some text removed from html files etc.
Then there's some work done on the interactive demo, mainly the top bar and stylsheet. The menu icon was moved from the right to the left because it ended up behind the close button on the featherweight popup on the website when anyone tried to preview the interactive demo. And I added a settings tab where the menu tab used to be and moved the wave logo to the middle. The Wave Logo, when clicked, hides the dropdown options (full screens the app).
There was also some funny stuff going on with the app preview images , with filter: invert(100%). Hopefully it's solved now.