Software Details Option
In the software section there's a link to the source code & docs of each application. Tonight a 3rd link was added to introduce a details page.  Which includes extra details about the software which is retrieved from GitHub. This info includes intensity of development, programming languages used, the latest updates/ comments (commits)  etc. There's just one issue now with parsing the changelog (atom). Basically the 3rd party service which turns atom into html expires in 30 days, that's no good. But with a website hosted in github (non php) it's not looking simple to parse atom to html client-side. But we're giving it a go