404 Page Fix
Say you are in a subdirectory and a files missing, the 404 will appear but with the subdirectory url, which makes the links in the 404 file reference the wrong source. So I put https://makeitwave.com/ after every href=" and src=" so that regardless of where and when the 404 page appears, it appears correctly. Sure there's a smarter way to solve this, but not now. Working on other stuff, this is just a quick fix. But it will need revisiting later on and sorted properly , because this is open source. so someone else might want to download the site, put on a different url (maybe even us one day) and this will be a bug in that later process. So it needs doing correctly, just not right now for me, that's a low low priority today.