Made the Website Open-Source (see the website footer)
In the footer there is now an option to follow website developement updates, download a copy of the website and see stats, like what languages are being used, how intense the development is etc. This is all fed from github directly to the website. The website is also hosted on github. Future plans here are to include a CMS for the website.