worked on some more documents
In the documents I have worked on the business plan.
There is now reference to 3 of the old business plans.
The text from the latest business plan is now in RST format. And can easily be converted to PDF. Although there is much work to do, to bring the text up to date.