introduced gaming (see interactive demo) & changed product design
Gaming will be a big part of Wave moving forward.
The virtual gamepad is ideal for advertising and cryptocurrency mining.
The virtual gamepad is also 4 player.
And the time spent on games almost doubles time spend browsing online and using the smart home control apps.
This should help ensure the revenue per unit exceeds the amount required to pay the ISP's.
The product design is now based on autopi.io - including a CAT1 Modem on a Raspberry Pi Zero