mining
the jse mining doesn't work in the offline UI.
so we've embedded an iframe to a page on the website which contains the mining script. This way JSEcoin and the law is happy because end-users are effectively viewing a webpage, when in fact, it's in a tiny iframe in the wave dashboard. So when someone is gaming, home controlling, viewing ip cameras or media etc (via the dashboard) crypto is being mined.