mining bug
so crypto coins don't mine with the workaround theorised in the last update. So we're trying it the other way around. We're making the local dashboard look at the mining webpage, then within the webpage it'll display the local dashboard within an iframe. This will mean the device will have to be online for the dashboard to work, but a script can be written another day to say if the device is offline then redirect to an offline dashboard without the mining. and/or employ offline mining if possible. Another problem for another day