taking nonsense elsewhere
we're moving this test over to another url. Since it's beginning to pollute the changelog of our main website. point to note for future. Since the changelog is visible to the public now, via our website and we don't want to deter investors and followers with our bullsh*t trial and error test data and logs