moving test with m3u and iptv to blog server
the log for these changes are public, via the website.
we don't want to congest them with testing. Only milestones and major developments.
so we'll use the blog url in future, since the change logs for the blog aren't anywhere but github. whereas the website and WaveOS have published their changelogs to the webpage and is easily readible by followers of this project.