removed the energy aspect
The energy management side of life is awesome, with apps and technology like sense and sensibo. But these solutions aren't really ready to be brought into Wave without unfeasible amounts of work e.g. Wave is a multi-user system with self-hosted html5 webapps which work online and offline. The fashion right now with technology is to ransom end-users into sending their data to the cloud, then using an app they can purchase that data back in pretty layouts. And integration also depends on api's through the cloud. This isn't consistent with Eric Schmit''s technology law of "social, mobile and local", not really. Wave is really social mobile and local at every level.  The nearest we got to energy management in Wave was emoncms, but they made each channel you monitor something you have to purchase. This is insanity to us here at Wave - bloody buying back your own data.  But it's a trend and people fall for it. At least with Wave everything in an end-users home that talks to his/her Wave app (locally or remotely) is free.  For us it would be as unethical as entering someone's home and putting a toll booth on their access to the bathroom.