Major Changes - Demo, Packages, Software & docs
Contact Us page addresses updated.  There are now 3 seperate demo's - one for mobile devices, one for laptops & tablets and one for the TV (HDMI out).  The website visitor can select one of the three prior to launch to see the differences.  The demo select page presents an opportunity to show the website visitor a typical configuration and the extrodinary capabilities of Wave in the home.  Including local and remote access and multiple connected devices.  The software page has been updated to show the specific hardware Wave has been tested and working on,  so there is no misunderstandings about what Single board computers WaveOS can operate. The website blog and website is also scrapping further development of version 0.5 and focusing on version 0.6, which will revisit and redevelop the whole solution from scratch using the lessons learn up until the most recent development (0.5.8).  version 0.5 is the result of the seed funding (WRT) and the low interest loans  (10% APR / WIT) and so 0.5 (soon to become version 0.6 branch a) will continue forward as the 3G/HSPA+ (SD IPTV) Solution - with some work it will become purposeful for a demostration unit only (not scalable, not for beta testing nor retail).  The 4G/LTE (HD IPTV) solution e.g. 0.6(branch b), will be developed using branch a as a guideline and with higher priority and financed using WIN Token Funds and Series A Funding.