Update to Document Library - WRT & CBT Scheme
CBT Scheme 0.0.2 -  Added the account balances to the document. Removed ability to withdraw gratuity on Royalty Tokens. Initial investment + interest only!. Closing date for submitting Revolut balances to Wave has changed from the 18th to the 18th August 2019.

WRT Tokens 0.1.4 - Updated to include interest to date, a feature introduced in April 2019 and Calculated for the CBT Scheme and affiliation with banking application Revolut